TrainWorkflow/requirements.txt
adal==1.2.6
azure-common==1.1.26
azure-core==1.12.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.1.0
azure-mgmt-storage==11.2.0
azureml-core==1.24.0.post1
azureml-dataprep==2.11.1
azureml-dataprep-native==30.0.0
azureml-dataprep-rslex==1.9.0
azureml-dataset-runtime==1.24.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.2
distro==1.5.0
docker==4.4.4
dotnetcore2==2.1.20
idna==2.10
importlib-metadata==3.7.2
isodate==0.6.0
jeepney==0.6.0
jmespath==0.10.0
jsonpickle==2.0.0
msal==1.10.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
pathspec==0.8.1
portalocker==1.7.1
pyarrow==1.0.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
SecretStorage==3.3.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.58.0
zipp==3.4.1
